Separate server code into an abstract class plus a basic and a full featured server class
